requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
absl-py==0.11.0           # via tensorboard, tensorflow
argon2-cffi==20.1.0       # via notebook
astunparse==1.6.3         # via tensorflow
async-generator==1.10     # via nbclient
attrs==20.3.0             # via jsonschema, pytest
backcall==0.2.0           # via ipython
bleach==3.2.1             # via nbconvert
cachetools==4.1.1         # via google-auth
certifi==2020.11.8        # via requests
cffi==1.14.4              # via argon2-cffi
chardet==3.0.4            # via requests
click==7.1.2              # via pip-tools
decorator==4.4.2          # via ipython
defusedxml==0.6.0         # via nbconvert
entrypoints==0.3          # via nbconvert
gast==0.3.3               # via tensorflow
google-auth-oauthlib==0.4.2  # via tensorboard
google-auth==1.23.0       # via google-auth-oauthlib, tensorboard
google-pasta==0.2.0       # via tensorflow
grpcio==1.33.2            # via tensorboard, tensorflow
h5py==2.10.0              # via tensorflow
idna==2.10                # via requests
iniconfig==1.1.1          # via pytest
ipykernel==5.3.4          # via ipywidgets, jupyter, jupyter-console, notebook, qtconsole
ipython-genutils==0.2.0   # via nbformat, notebook, qtconsole, traitlets
ipython==7.19.0           # via ipykernel, ipywidgets, jupyter-console
ipywidgets==7.5.1         # via jupyter
isort==5.6.4              # via -r requirements.in
jedi==0.17.2              # via ipython
jinja2==2.11.2            # via nbconvert, notebook
jsonschema==3.2.0         # via nbformat
jupyter-client==6.1.7     # via ipykernel, jupyter-console, nbclient, notebook, qtconsole
jupyter-console==6.2.0    # via jupyter
jupyter-core==4.7.0       # via jupyter-client, nbconvert, nbformat, notebook, qtconsole
jupyter==1.0.0            # via -r requirements.in
jupyterlab-pygments==0.1.2  # via nbconvert
keras-preprocessing==1.1.2  # via tensorflow
markdown==3.3.3           # via tensorboard
markupsafe==1.1.1         # via jinja2
mistune==0.8.4            # via nbconvert
mypy-extensions==0.4.3    # via mypy
mypy==0.790               # via -r requirements.in
nbclient==0.5.1           # via nbconvert
nbconvert==6.0.7          # via jupyter, notebook
nbformat==5.0.8           # via ipywidgets, nbclient, nbconvert, notebook
nest-asyncio==1.4.3       # via nbclient
notebook==6.1.5           # via jupyter, widgetsnbextension
numpy==1.18.5             # via -r requirements.in, h5py, keras-preprocessing, opt-einsum, tensorboard, tensorflow
oauthlib==3.1.0           # via requests-oauthlib
opt-einsum==3.3.0         # via tensorflow
packaging==20.7           # via bleach, pytest
pandocfilters==1.4.3      # via nbconvert
parso==0.7.1              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==5.4.0          # via -r requirements.in
pluggy==0.13.1            # via pytest
prometheus-client==0.9.0  # via notebook
prompt-toolkit==3.0.8     # via ipython, jupyter-console
protobuf==3.14.0          # via tensorboard, tensorflow
ptyprocess==0.6.0         # via pexpect, terminado
py==1.9.0                 # via pytest
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pycparser==2.20           # via cffi
pygments==2.7.2           # via ipython, jupyter-console, jupyterlab-pygments, nbconvert, qtconsole
pyhamcrest==2.0.2         # via -r requirements.in
pyparsing==2.4.7          # via packaging
pyrsistent==0.17.3        # via jsonschema
pytest==6.1.2             # via -r requirements.in
python-dateutil==2.8.1    # via jupyter-client
pyzmq==20.0.0             # via jupyter-client, notebook, qtconsole
qtconsole==5.0.1          # via jupyter
qtpy==1.9.0               # via qtconsole
requests-oauthlib==1.3.0  # via google-auth-oauthlib
requests==2.25.0          # via requests-oauthlib, tensorboard
rsa==4.6                  # via google-auth
send2trash==1.5.0         # via notebook
six==1.15.0               # via absl-py, argon2-cffi, astunparse, bleach, google-auth, google-pasta, grpcio, h5py, jsonschema, keras-preprocessing, pip-tools, protobuf, python-dateutil, tensorboard, tensorflow
tensorboard-plugin-wit==1.7.0  # via tensorboard
tensorboard==2.4.0        # via tensorflow
tensorflow-estimator==2.3.0  # via tensorflow
tensorflow==2.3.1         # via -r requirements.in
termcolor==1.1.0          # via tensorflow
terminado==0.9.1          # via notebook
testpath==0.4.4           # via nbconvert
toml==0.10.2              # via pytest
tornado==6.1              # via ipykernel, jupyter-client, notebook, terminado
traitlets==5.0.5          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbclient, nbconvert, nbformat, notebook, qtconsole
typed-ast==1.4.1          # via mypy
typing-extensions==3.7.4.3  # via mypy
urllib3==1.26.2           # via requests
wcwidth==0.2.5            # via prompt-toolkit
webencodings==0.5.1       # via bleach
werkzeug==1.0.1           # via tensorboard
wheel==0.35.1             # via astunparse, tensorboard, tensorflow
widgetsnbextension==3.5.1  # via ipywidgets
wrapt==1.12.1             # via tensorflow
yapf==0.30.0              # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
